dbt cloud support
Closes #98
