+++ packages/common/src/index.ts
@@ -364,6 +364,62 @@ export const friendlyName = (text: string): string => {
     return [capitalize(first), ...rest].join(' ')
 }
 
+
+// DBT CONFIG
+export type DbtNode = {
+    unique_id: string
+    resource_type: string,
+}
+export type DbtModelNode = DbtNode & {
+    columns: { [name: string]: DbtModelColumn },
+    meta: DbtModelMetadata,
+    database: string,
+    schema: string,
+    name: string,
+    relation_name: string,
+    depends_on: DbtTableDependency,
+    description?: string,
+}
+type DbtTableDependency = {
+    nodes: string[]
+}
+export type DbtModelColumn = {
+    name: string,
+    description?: string,
+    meta: DbtColumnMetadata,
+    data_type?: string,
+}
+
+
+// CUSTOM LIGHTDASH CONFIG IN DBT
+type DbtModelMetadata = DbtModelLightdashConfig & {}
+
+type DbtModelLightdashConfig = {
+    joins?: DbtModelJoin[]
+}
+type DbtModelJoin = {
+    join: string,
+    sql_on: string,
+}
+type DbtColumnMetadata = DbtColumnLightdashConfig & {}
+type DbtColumnLightdashConfig = {
+    dimension?: DbtColumnLightdashDimension,
+    metrics?: {[metricName: string]: DbtColumnLightdashMetric}
+}
+
+type DbtColumnLightdashDimension = {
+    name?: string,
+    type?: DimensionType,
+    description?: string,
+    sql?: string,
+}
+
+type DbtColumnLightdashMetric = {
+    type: MetricType,
+    description?: string,
+    sql?: string,
+}
+
 type ApiErrorDetail = {
     name: string,
     statusCode: number,
@@ -436,3 +492,32 @@ export type ApiResponse =
     | ApiTableResponse
     | ApiStatusResponse
     | ApiRefreshResponse
+
+export interface DbtCatalogNode {
+    metadata: DbtCatalogNodeMetadata;
+    columns: {
+        [k: string]: DbtCatalogNodeColumn;
+    };
+}
+
+export interface DbtCatalogNodeMetadata {
+    type: string;
+    database: string;
+    schema: string;
+    name: string;
+    comment?: string;
+    owner?: string;
+}
+
+export interface DbtCatalogNodeColumn {
+    type: string;
+    comment?: string;
+    index: number;
+    name: string;
+}
+
+export interface DbtCatalog {
+    nodes: {
+        [k: string]: DbtCatalogNode;
+    };
+}
